Squashed Merge V2 -> Main

+++ Benchmarks/Deterministic.GameFramework.Physics.Benchmarks/PhysicsSystemBenchmark.cs
@@ -111,7 +111,7 @@ public class PhysicsSystemBenchmark
 public class FakeGameTime : IGameTime
 {
     public long CurrentTick { get; set; }
-    public float FixedDeltaTime { get; set; } = 1.0f / 60.0f;
+    public Float FixedDeltaTime { get; set; } = (Float)1 / (Float)60;
     public int TickRate { get; set; } = 60;
     public bool IsResimulating { get; set; } = false;
 }
